Add files via upload
Code in Python 3 that led to the user to create a List of integers and finds all the tuples of elements that sums a specific number that the user also imputs for the comparison. The code sort the list to find the tuples in an scallable cycle that iterates in O(n/2) of the size of the list.